SECURITY: Properly escape HTML in poll tooltips
This fixes a severe XSS vulnerability.

Bootstrap tooltips display the content of an element's data-title and
data-content attributes, which may contain escaped HTML that is
then unescaped. Therefore, unsafe content must be escaped *twice*
when inserted into these tooltips. This can be done using the
jinja2 filter 'forceescape' (the normal 'e' filter is not sufficient,
as it will escape the HTML only once).